[OC-77] proxy: refactor proxy settings code
Move platform-agnostic proxy code to own classes - this is needed to
accommodate Windows proxy implementation.
